ML_requirements.txt
scipy
numpy
matplotlib
pandas
sklearn
tensorflow-gpu
keras
ipython
six
protobuf==3.1.0
wheel
jupyter
mnist
